refactor: Input/Output managers

+++ src/py_wav/io/pyaudio.py
@@ -1,27 +1,34 @@
 from __future__ import annotations
 
-from typing import Optional, ContextManager
-
 import numpy as np
 import pyaudio
 
 from custom_types import Hz, Frames
-from py_wav.io.streaming import StreamChunk, ChunkMetadata, AudioChunkStream, MetadataChunkStream, ChunkIO, \
-    MetadataTimer
+from py_wav.io.streaming import StreamChunk, InputManager, OutputManager, IOContext
 
 
-class PyAudioContext(ContextManager):
-    def __init__(self, fs: Hz, channels: int, chunk_size: Frames,
-                 do_input: bool, do_output: bool):
+class PyAudioContext(IOContext):
+    def __init__(self, fs: Hz, channels: int, chunk_size: Frames):
         self.chunk_size = chunk_size
         self.fs = int(fs)
         self.channels = int(channels)
-        self.do_input = do_input
-        self.do_output = do_output
+        self.do_input = False
+        self.do_output = False
         self.p = None
         self.stream = None
+        self.counter = 0
+
+    def use_for_input(self):
+        self.do_input = True
+
+    def use_for_output(self):
+        self.do_output = True
 
     def __enter__(self):
+        self.counter += 1
+        if self.counter > 1:
+            return self
+        print("starting up pyaudio")
         self.p = pyaudio.PyAudio()
         self.stream = self.p.open(format=pyaudio.paFloat32,
                                   channels=self.channels,
@@ -33,17 +40,25 @@ class PyAudioContext(ContextManager):
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
+        self.counter -= 1
+        if self.counter > 0:
+            return
+        print("shutting down pyaudio")
         self.stream.stop_stream()
         self.stream.close()
         self.p.terminate()
 
 
-class PyAudioStreaming(ChunkIO[PyAudioContext]):
+class PyAudioStreaming(InputManager, OutputManager):
     def __init__(self, fs: Hz, channels: int, chunk_size: Frames):
-        ChunkIO.__init__(self, fs, chunk_size, name="pyaudio")
         self.chunk_size = chunk_size
         self.fs = int(fs)
         self.channels = int(channels)
+        self.context = PyAudioContext(
+            fs=self.fs,
+            channels=self.channels,
+            chunk_size=self.chunk_size,
+        )
 
     def read_input(self, ctx: PyAudioContext):
         in_data = ctx.stream.read(self.chunk_size)
@@ -53,11 +68,5 @@ class PyAudioStreaming(ChunkIO[PyAudioContext]):
         frames = chunk.buf.astype(np.float32).tobytes()
         ctx.stream.write(frames, num_frames=len(chunk.buf))
 
-    def streaming_context(self, do_input: bool, do_output: bool) -> PyAudioContext:
-        return PyAudioContext(
-            fs=self.fs,
-            channels=self.channels,
-            chunk_size=self.chunk_size,
-            do_input=do_input,
-            do_output=do_output,
-        )
+    def get_context(self) -> PyAudioContext:
+        return self.context
